fix: 업로드 성공시 메인 이동 에러 수정 업로드 성공시 메인 이동 에러 수정 및, api 파일 분리
