+++ src/apis/main.ts
@@ -1,9 +1,9 @@
+import { PostData } from '@models/postData';
 import http from './http';
 
 export const getMarkers = () => {
   return http.get('/cats');
 };
-
 
 export const getCatDetail = (id: number) => {
   return http.get(`/cats/${id}`);
@@ -11,4 +11,19 @@ export const getCatDetail = (id: number) => {
 
 export const postLike = (id: number) => {
   return http.post('/like', { id });
+};
+
+export const postCat = (form: PostData) => {
+  const formData = new FormData();
+  // const location = form.address.split(' ').slice(0, 2).join(' ');
+  formData.append('catImage', form.imageUrl!);
+  formData.append('description', form.description);
+  formData.append('latitude', `${form.position.latitude}`);
+  formData.append('longitude', `${form.position.longitude}`);
+  formData.append('location', 'gannamgu');
+  return http.post('/cats', formData, {
+    headers: {
+      'Content-Type': 'multipart/form-data',
+    },
+  });
 };
